# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yoga Studio Finance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Customer {
  id: number;
  fullName: string;
  phone?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCustomer {
  fullName: string;
  phone?: string | null;
  notes?: string | null;
}

export interface Product {
  id: number;
  name: string;
  /** Price in cents */
  defaultPrice: number;
  active: boolean;
  sortOrder: number;
}

export interface CreateProduct {
  name: string;
  /** Price in cents */
  defaultPrice: number;
  active?: boolean;
  sortOrder?: number;
}

export type PaymentPaymentMethod =
  (typeof PaymentPaymentMethod)[keyof typeof PaymentPaymentMethod];

export const PaymentPaymentMethod = {
  contanti: "contanti",
  pos: "pos",
  bonifico: "bonifico",
} as const;

export interface Payment {
  id: number;
  customerId: number;
  productId: number;
  /** Amount in cents */
  amount: number;
  paymentMethod: PaymentPaymentMethod;
  date: string;
  note?: string | null;
  createdAt: string;
  updatedAt: string;
  customerName: string;
  productName: string;
}

export type CreatePaymentPaymentMethod =
  (typeof CreatePaymentPaymentMethod)[keyof typeof CreatePaymentPaymentMethod];

export const CreatePaymentPaymentMethod = {
  contanti: "contanti",
  pos: "pos",
  bonifico: "bonifico",
} as const;

export interface CreatePayment {
  customerId: number;
  productId: number;
  /** Amount in cents */
  amount: number;
  paymentMethod: CreatePaymentPaymentMethod;
  date: string;
  note?: string | null;
}

export type TeacherCompensationType =
  (typeof TeacherCompensationType)[keyof typeof TeacherCompensationType];

export const TeacherCompensationType = {
  hourly: "hourly",
  manual: "manual",
} as const;

export interface Teacher {
  id: number;
  name: string;
  compensationType: TeacherCompensationType;
  /** Rate in cents */
  hourlyRate?: number | null;
  active: boolean;
}

export type CreateTeacherCompensationType =
  (typeof CreateTeacherCompensationType)[keyof typeof CreateTeacherCompensationType];

export const CreateTeacherCompensationType = {
  hourly: "hourly",
  manual: "manual",
} as const;

export interface CreateTeacher {
  name: string;
  compensationType: CreateTeacherCompensationType;
  /** Rate in cents */
  hourlyRate?: number | null;
  active?: boolean;
}

export interface TeacherMonthlyHours {
  id: number;
  teacherId: number;
  month: string;
  hoursWorked: number;
  /** Manual cost in cents for manual compensation type */
  manualCost?: number | null;
}

export interface UpsertTeacherHours {
  month: string;
  hoursWorked: number;
  /** Manual cost in cents */
  manualCost?: number | null;
}

export interface OtherCost {
  id: number;
  month: string;
  category: string;
  /** Amount in cents */
  amount: number;
  note?: string | null;
}

export interface CreateOtherCost {
  month: string;
  category: string;
  /** Amount in cents */
  amount: number;
  note?: string | null;
}

export interface MonthlySummary {
  month: string;
  /** Total revenue in cents */
  revenue: number;
  /** Total teacher costs in cents */
  teacherCosts: number;
  /** Total other costs in cents */
  otherCosts: number;
  /** Estimated taxes in cents */
  estimatedTaxes: number;
  /** Net profit in cents */
  netProfit: number;
}

export interface TeacherAnalysis {
  teacherId: number;
  teacherName: string;
  hoursWorked: number;
  /** In cents */
  totalCost: number;
  /** In cents */
  avgHourlyValue: number;
  /** In cents */
  estimatedRevenue: number;
  /** In cents */
  estimatedMargin: number;
}

export interface TaxSettings {
  id: number;
  /** Tax rate as percentage (e.g. 22 for 22%) */
  taxRate: number;
}

export interface UpdateTaxSettings {
  taxRate: number;
}

export interface DefaultCost {
  category: string;
  /** Amount in cents */
  amount: number;
}

export type ListCustomersParams = {
  search?: string;
};

export type ListPaymentsParams = {
  /**
   * Month in YYYY-MM format
   */
  month?: string;
};

export type GetTeacherHoursParams = {
  /**
   * Month in YYYY-MM format
   */
  month: string;
};

export type ListOtherCostsParams = {
  /**
   * Month in YYYY-MM format
   */
  month: string;
};