# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Yoga Studio Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomer"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments with optional month filter
      parameters:
        - name: month
          in: query
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayment"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    put:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayment"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /other-costs:
    get:
      operationId: listOtherCosts
      tags: [otherCosts]
      summary: List other costs for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: List of other costs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OtherCost"
    post:
      operationId: createOtherCost
      tags: [otherCosts]
      summary: Create an other cost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOtherCost"
      responses:
        "201":
          description: Created other cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtherCost"
  /other-costs/{id}:
    put:
      operationId: updateOtherCost
      tags: [otherCosts]
      summary: Update an other cost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOtherCost"
      responses:
        "200":
          description: Updated other cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtherCost"
    delete:
      operationId: deleteOtherCost
      tags: [otherCosts]
      summary: Delete an other cost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /settings/tax:
    get:
      operationId: getTaxSettings
      tags: [settings]
      summary: Get tax settings
      responses:
        "200":
          description: Tax settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxSettings"
    put:
      operationId: updateTaxSettings
      tags: [settings]
      summary: Update tax settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaxSettings"
      responses:
        "200":
          description: Updated tax settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxSettings"
  /settings/default-costs:
    get:
      operationId: getDefaultCosts
      tags: [settings]
      summary: Get default recurring costs
      responses:
        "200":
          description: Default costs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefaultCost"
    put:
      operationId: updateDefaultCosts
      tags: [settings]
      summary: Update default recurring costs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/DefaultCost"
      responses:
        "200":
          description: Updated default costs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DefaultCost"
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - createdAt
        - updatedAt
    CreateCustomer:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - fullName
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        defaultPrice:
          type: integer
          description: Price in cents
        active:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - name
        - defaultPrice
        - active
        - sortOrder
    CreateProduct:
      type: object
      properties:
        name:
          type: string
        defaultPrice:
          type: integer
          description: Price in cents
        active:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
        - defaultPrice
    Payment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        productId:
          type: integer
        amount:
          type: integer
          description: Amount in cents
        paymentMethod:
          type: string
          enum: [contanti, pos, bonifico]
        date:
          type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        customerName:
          type: string
        productName:
          type: string
      required:
        - id
        - customerId
        - productId
        - amount
        - paymentMethod
        - date
        - createdAt
        - updatedAt
        - customerName
        - productName
    CreatePayment:
      type: object
      properties:
        customerId:
          type: integer
        productId:
          type: integer
        amount:
          type: integer
          description: Amount in cents
        paymentMethod:
          type: string
          enum: [contanti, pos, bonifico]
        date:
          type: string
        note:
          type: string
          nullable: true
      required:
        - customerId
        - productId
        - amount
        - paymentMethod
        - date
...
    OtherCost:
      type: object
      properties:
        id:
          type: integer
        month:
          type: string
        category:
          type: string
        amount:
          type: integer
          description: Amount in cents
        note:
          type: string
          nullable: true
      required:
        - id
        - month
        - category
        - amount
    CreateOtherCost:
      type: object
      properties:
        month:
          type: string
        category:
          type: string
        amount:
          type: integer
          description: Amount in cents
        note:
          type: string
          nullable: true
      required:
        - month
        - category
        - amount
...
    TaxSettings:
      type: object
      properties:
        id:
          type: integer
        taxRate:
          type: number
          description: Tax rate as percentage (e.g. 22 for 22%)
      required:
        - id
        - taxRate
    UpdateTaxSettings:
      type: object
      properties:
        taxRate:
          type: number
      required:
        - taxRate
    DefaultCost:
      type: object
      properties:
        category:
          type: string
        amount:
          type: integer
          description: Amount in cents
      required:
        - category
        - amount