# lib/api-zod/src/generated/types/createCustomer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yoga Studio Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCustomer {
  fullName: string;
  phone?: string | null;
  notes?: string | null;
}