# lib/api-zod/src/generated/types/createProduct.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yoga Studio Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProduct {
  name: string;
  /** Price in cents */
  defaultPrice: number;
  active?: boolean;
  sortOrder?: number;
}