# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yoga Studio Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentPaymentMethod } from "./paymentPaymentMethod";

export interface Payment {
  id: number;
  customerId: number;
  productId: number;
  /** Amount in cents */
  amount: number;
  paymentMethod: PaymentPaymentMethod;
  date: string;
  note?: string | null;
  createdAt: string;
  updatedAt: string;
  customerName: string;
  productName: string;
}