# lib/api-zod/src/generated/types/taxSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Yoga Studio Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface TaxSettings {
  id: number;
  /** Tax rate as percentage (e.g. 22 for 22%) */
  taxRate: number;
}